requirements.txt
PyQt5
flask
flask-socketio
eventlet
python-socketio
websocket-client
requests
